requirements.txt
autoattack==0.1
datasets==3.5.0
detectron2.egg==info
einops==0.8.1
ftfy==6.3.1
h5py==3.13.0
matplotlib==3.10.6
nltk==3.9.1
numpy==2.3.3
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
openpyxl==3.1.5
pandas==2.3.2
Pillow==11.3.0
pyarrow==19.0.1
pycountry==24.6.1
regex==2024.11.6
scikit_learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==75.8.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
